chore: add ssr only to times components article
