+++ packages/article-skeleton/src/article-body/article-body.js
@@ -4,6 +4,7 @@ import PropTypes from "prop-types";
 import { AdContainer } from "@times-components/ad";
 import LazyLoad from "@times-components/lazy-load";
 import ArticleImage from "@times-components/article-image";
+import StaticContent from '../static-content';
 
 import ArticleParagraph, {
   DropCapView
@@ -64,43 +65,6 @@ import {
 } from "../styles/article-body/responsive";
 
 const deckApiUrl = "https://gobble.timesdev.tools/deck/api/deck-post-action/";
-
-function useStaticContent() {
-    const ref = useRef(null);
-    const [render, setRender] = useState(typeof window === 'undefined');
-
-    useEffect(() => {
-        const isEmpty = ref.current.innerHTML === '';
-        if (isEmpty) {
-            setRender(true);
-        }
-    }, []);
-
-    return [render, ref];
-}
-
-export default function StaticContent({
-    children,
-    element = 'div',
-    html,
-    ...props
-}) {
-    const [shouldRender, ref] = useStaticContent();
-
-    if (shouldRender) {
-        return createElement(element, {
-            ...props,
-            children,
-        });
-    }
-
-    return createElement(element, {
-        ...props,
-        ref,
-        suppressHydrationWarning: true,
-        dangerouslySetInnerHTML: { __html: html },
-    });
-}
 
 export const responsiveDisplayWrapper = displayType => {
   switch (displayType) {
@@ -524,13 +488,11 @@ const renderers = ({
   },
   paywall(key, attributes, children) {
     return (
-        <div>
-      <span className={paidContentClassName} key={key}>
-        {children}
-      </span>
-      <StaticContent
-            html={"<div id='zephr-target-pw'>Here's static emptiness</div>"}/>
-        </div>
+      <div>
+        <span className={paidContentClassName} key={key}>
+          {children}
+        </span>
+      </div>
     );
   },
   pullQuote(
